naming schema: hazelcast
