+++ dd-java-agent/instrumentation/hazelcast-3.9/src/main/java/datadog/trace/instrumentation/hazelcast39/ClientInvocationDecorator.java
@@ -6,19 +6,24 @@ import static datadog.trace.instrumentation.hazelcast39.HazelcastConstants.HAZEL
 import static datadog.trace.instrumentation.hazelcast39.HazelcastConstants.HAZELCAST_NAME;
 import static datadog.trace.instrumentation.hazelcast39.HazelcastConstants.HAZELCAST_OPERATION;
 import static datadog.trace.instrumentation.hazelcast39.HazelcastConstants.HAZELCAST_SERVICE;
+import static datadog.trace.instrumentation.hazelcast39.HazelcastConstants.INSTRUMENTATION_NAME;
 
+import datadog.trace.api.Config;
+import datadog.trace.api.naming.SpanNaming;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import datadog.trace.bootstrap.instrumentation.api.InternalSpanTypes;
 import datadog.trace.bootstrap.instrumentation.api.UTF8BytesString;
 import datadog.trace.bootstrap.instrumentation.decorator.ClientDecorator;
 import datadog.trace.util.Strings;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /** Decorate Hazelcast client invocations with relevant contextual information. */
 public class ClientInvocationDecorator extends ClientDecorator {
 
-  private static final Logger log = LoggerFactory.getLogger(ClientInvocationDecorator.class);
+  private static final String SERVICE_NAME =
+      SpanNaming.instance()
+          .namingSchema()
+          .cache()
+          .service(Config.get().getServiceName(), INSTRUMENTATION_NAME);
 
   public static final ClientInvocationDecorator DECORATE = new ClientInvocationDecorator();
 
@@ -39,7 +44,7 @@ public class ClientInvocationDecorator extends ClientDecorator {
 
   @Override
   protected String service() {
-    return COMPONENT_NAME.toString();
+    return SERVICE_NAME;
   }
 
   /** Decorate trace based on service execution metadata. */
@@ -67,15 +72,12 @@ public class ClientInvocationDecorator extends ClientDecorator {
   }
 
   public AgentSpan onHazelcastInstance(final AgentSpan span, String instanceName) {
-
     span.setTag(HAZELCAST_INSTANCE, instanceName);
-
     return span;
   }
 
   /** Annotate the span with the results of the operation. */
   public AgentSpan onResult(final AgentSpan span, Object result) {
-
     return span;
   }
 }
