+++ dd-java-agent/instrumentation/hazelcast-3.9/src/test/groovy/test/hazelcast/v39/HazelcastTest.groovy
@@ -1,10 +1,23 @@
 package test.hazelcast.v39
 
+import com.hazelcast.client.HazelcastClient
+import com.hazelcast.client.config.ClientConfig
+import com.hazelcast.config.Config
+import com.hazelcast.core.Hazelcast
+import com.hazelcast.core.HazelcastInstance
 import com.hazelcast.core.Message
 import com.hazelcast.core.MessageListener
 import com.hazelcast.query.Predicates
+import datadog.trace.agent.test.asserts.ListWriterAssert
+import datadog.trace.agent.test.asserts.TraceAssert
+import datadog.trace.agent.test.naming.VersionedNamingTestBase
+import datadog.trace.agent.test.utils.PortUtils
 import datadog.trace.api.DDSpanTypes
 import datadog.trace.bootstrap.instrumentation.api.Tags
+import datadog.trace.common.writer.ListWriter
+import datadog.trace.core.DDSpan
+import net.bytebuddy.utility.RandomString
+import spock.lang.Shared
 import spock.util.concurrent.BlockingVariable
 
 import java.util.concurrent.TimeUnit
@@ -12,7 +25,64 @@ import java.util.concurrent.TimeUnit
 import static datadog.trace.agent.test.utils.TraceUtils.basicSpan
 import static datadog.trace.agent.test.utils.TraceUtils.runUnderTrace
 
-class HazelcastTest extends AbstractHazelcastTest {
+abstract class HazelcastTest extends VersionedNamingTestBase {
+
+  @Shared
+  String randomName
+  @Shared
+  HazelcastInstance h1, client
+
+  final resourceNamePattern = ~/^(?<operation>(?<service>[A-Z]\w+)\.[a-z]\w+)(?: (?<name>.+))?$/
+
+  /** Filter our Client operations. They can happen at seemingly random times and yield inconsistent test results. */
+  final ListWriter.Filter defaultFilter = new ListWriter.Filter() {
+    @Override
+    boolean accept(List<DDSpan> trace) {
+      return !(trace.size() == 1 && trace.get(0).getResourceName().startsWithAny("Client."))
+    }
+  }
+
+  @Override
+  protected void configurePreAgent() {
+    super.configurePreAgent()
+
+    injectSysConfig("dd.integration.hazelcast.enabled", "true")
+  }
+
+  @Override
+  def setupSpec() {
+    def port = PortUtils.randomOpenPort()
+    def groupName = RandomString.make(8)
+    def groupPassword = RandomString.make(8)
+
+    def serverConfig = new Config()
+    serverConfig.groupConfig.name = groupName
+    serverConfig.groupConfig.password = groupPassword
+
+    serverConfig.networkConfig.port = port
+    serverConfig.networkConfig.portAutoIncrement = false
+    serverConfig.networkConfig.join.multicastConfig.enabled = false
+    configureServer(serverConfig)
+
+    h1 = Hazelcast.newHazelcastInstance(serverConfig)
+
+    def clientConfig = new ClientConfig()
+    clientConfig.groupConfig.name = groupName
+    clientConfig.groupConfig.password = groupPassword
+    clientConfig.networkConfig.addAddress("127.0.0.1:$port")
+    client = HazelcastClient.newHazelcastClient(clientConfig)
+  }
+
+  @Override
+  def cleanupSpec() {
+    Hazelcast.shutdownAll()
+  }
+
+  def setup() {
+    TEST_WRITER.setFilter(defaultFilter)
+    randomName = randomResourceName()
+  }
+
 
   def "map"() {
     setup:
@@ -253,8 +323,8 @@ class HazelcastTest extends AbstractHazelcastTest {
     assertTraces(1) {
       trace(1) {
         span {
-          serviceName "hazelcast-sdk"
-          operationName "hazelcast.invoke"
+          serviceName service()
+          operationName operation()
           resourceName "List.get $randomName"
           parent()
           spanType DDSpanTypes.HTTP_CLIENT
@@ -295,5 +365,82 @@ class HazelcastTest extends AbstractHazelcastTest {
 
     cleanup:
     serverLock?.destroy()
+  }
+
+  void hazelcastTrace(ListWriterAssert writer, String name) {
+    writer.trace(1) {
+      hazelcastSpan(it, name)
+    }
+  }
+
+  def hazelcastSpan(TraceAssert trace, String name, boolean isParent = true) {
+    def matcher = name =~ resourceNamePattern
+    assert matcher.matches()
+
+    trace.span {
+      serviceName service()
+      resourceName name
+      operationName operation()
+      spanType DDSpanTypes.HTTP_CLIENT
+      errored false
+      measured true
+      if (isParent) {
+        parent()
+      } else {
+        childOfPrevious()
+      }
+      tags {
+        "$Tags.COMPONENT" "hazelcast-sdk"
+        "$Tags.SPAN_KIND" Tags.SPAN_KIND_CLIENT
+        "hazelcast.name" matcher.group("name")
+        "hazelcast.operation" matcher.group("operation")
+        "hazelcast.service" matcher.group("service")
+        "hazelcast.instance" client.name
+        "hazelcast.correlationId" Long
+        defaultTags()
+      }
+    }
+  }
+
+  void configureServer(Config config) {}
+
+  def randomResourceName(int length = 8) {
+    RandomString.make(length)
+  }
+}
+
+class HazelcastV0ForkedTest extends HazelcastTest {
+
+  @Override
+  protected int version() {
+    return 0
+  }
+
+  @Override
+  protected String service() {
+    return "hazelcast-sdk"
+  }
+
+  @Override
+  protected String operation() {
+    return "hazelcast.invoke"
+  }
+}
+
+class HazelcastV1ForkedTest extends HazelcastTest {
+
+  @Override
+  protected int version() {
+    return 1
+  }
+
+  @Override
+  protected String service() {
+    return datadog.trace.api.Config.get().getServiceName() + "-hazelcast"
+  }
+
+  @Override
+  protected String operation() {
+    return "hazelcast.command"
   }
 }
