+++ dd-java-agent/instrumentation/hazelcast-4.0/src/main/java/datadog/trace/instrumentation/hazelcast4/HazelcastConstants.java
@@ -1,5 +1,6 @@
 package datadog.trace.instrumentation.hazelcast4;
 
+import datadog.trace.api.naming.SpanNaming;
 import datadog.trace.bootstrap.instrumentation.api.UTF8BytesString;
 
 /** Holder for constant values shared across the instrumentation versions. */
@@ -15,7 +16,8 @@ public class HazelcastConstants {
   public static final String HAZELCAST_INSTANCE = "hazelcast.instance";
   public static final String HAZELCAST_CORRELATION_ID = "hazelcast.correlationId";
 
-  public static final CharSequence SPAN_NAME = UTF8BytesString.create("hazelcast.invoke");
-
+  public static final CharSequence SPAN_NAME =
+      UTF8BytesString.create(
+          SpanNaming.instance().namingSchema().cache().operation(INSTRUMENTATION_NAME));
   public static final CharSequence COMPONENT_NAME = UTF8BytesString.create("hazelcast-sdk");
 }
